clean old remnants

+++ Code/ww3d2/agg_def.h
@@ -35,8 +35,6 @@
 
 #if defined(_MSC_VER)
 #pragma once
-
-#include <cstdint>
 #endif
 
 #ifndef AGGREGATE_DEF_H
@@ -119,7 +117,7 @@ class AggregateDefClass
 		//
 		//	Public accessors
 		//
-		uint32_t							Class_ID (void) const					{ return m_MiscInfo.OriginalClassID; }
+		ULONG							Class_ID (void) const					{ return m_MiscInfo.OriginalClassID; }
 
 		//
 		//	Initialization
@@ -191,7 +189,7 @@ class AggregateDefClass
 		//
 		//	Private member data
 		//
-		uint32_t																m_Version;
+		DWORD																m_Version;
 		DynamicVectorClass<W3dAggregateSubobjectStruct *>	m_SubobjectList;
 		W3dAggregateInfoStruct										m_Info;
 		W3dAggregateMiscInfo											m_MiscInfo;
